CIL-1918 New email verification via "code"
An ACCESS Help Ticket was submitted containing the old "verify email
with link" email message. This implies that there is documentation out
there in the world which links directly to the old enrollment flows
rather than to https://operations.access-ci.org/identity/new-user . So I
modified the old enrollment flows to use the new
EmailVerificationEnroller wedge, and reverted the redirect links to the
old enrollment flows.